+++ app/src/main/java/com/thecode/infotify/framework/datasource/NewsApiRemoteServiceImpl.kt
@@ -5,25 +5,8 @@ import com.thecode.infotify.framework.datasource.network.model.NewsObjectRespons
 import com.thecode.infotify.framework.datasource.network.model.SourceObjectResponse
 import retrofit2.Response
 
-interface NewsApiRemoteService {
-    //region NEWS
-    suspend fun getAllNews(query: String, language: String, sortBy: String): NewsObjectResponse
 
-    suspend fun getTopHeadlinesBySources(sources: String): Response<NewsObjectResponse>
-
-    suspend fun getTopHeadlinesByCategory(category: String): NewsObjectResponse
-
-    suspend fun getAllNewsByQuery(query: String): Response<NewsObjectResponse>
-    //endregion
-
-    //region SOURCE
-    suspend fun getSources(): Response<SourceObjectResponse>
-
-    suspend fun getSourcesByLang(language: String): Response<SourceObjectResponse>
-    //endregion
-}
-
-class NewsApiRemoteServiceImpl constructor(
+class NewsApiRemoteServiceImpl(
     private val newsApi: NewsApi
 ) : NewsApiRemoteService {
